+++ src/Core/Migration/V6_7/Migration1742563555AddNotificationTable.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Administration\Migration\V6_4;
+namespace Shopware\Core\Migration\V6_7;
 
 use Doctrine\DBAL\Connection;
 use Shopware\Core\Framework\Log\Package;
@@ -8,15 +8,13 @@ use Shopware\Core\Framework\Migration\MigrationStep;
 
 /**
  * @internal
- *
- * @codeCoverageIgnore
  */
-#[Package('framework')]
-class Migration1632281097Notification extends MigrationStep
+#[Package('core')]
+class Migration1742563555AddNotificationTable extends MigrationStep
 {
     public function getCreationTimestamp(): int
     {
-        return 1632281097;
+        return 1742563555;
     }
 
     public function update(Connection $connection): void
@@ -25,7 +23,7 @@ class Migration1632281097Notification extends MigrationStep
             CREATE TABLE IF NOT EXISTS `notification` (
                 `id` BINARY(16) NOT NULL,
                 `status` VARCHAR(255) NOT NULL,
-                `message` VARCHAR(5000) NOT NULL,
+                `message` LONGTEXT NOT NULL,
                 `admin_only` tinyint(1) NOT NULL DEFAULT 0,
                 `required_privileges` json NULL,
                 `created_by_integration_id` BINARY(16) NULL,
@@ -37,10 +35,5 @@ class Migration1632281097Notification extends MigrationStep
                 CONSTRAINT `fk.notification.created_by_user_id` FOREIGN KEY (`created_by_user_id`) REFERENCES `user` (`id`) ON DELETE SET NULL ON UPDATE CASCADE
             ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
         ');
-    }
-
-    public function updateDestructive(Connection $connection): void
-    {
-        // implement update destructive
     }
 }
